requirements.txt
psycopg2-binary
psycopg2
python-dotenv
Flask
Flask-Cors
ratelimit
yfinance
pandas
langchain
langchain-google-genai
yflive
protobuf<3.21.0
apscheduler
gunicorn
flask-socketio
eventlet